ADD: Agregados policies

+++ app/Http/Controllers/PermissionController.php
@@ -11,6 +11,9 @@ class PermissionController extends Controller
 {
     public function index()
     {
+        // Policy: Llama al método 'viewAny' de la PermissionPolicy
+        $this->authorize('viewAny', Permission::class);
+
         $permissions = Permission::all();
 
         return PermissionResource::collection($permissions)->response()->setStatusCode(200);
@@ -18,11 +21,17 @@ class PermissionController extends Controller
 
     public function show(Permission $permission)
     {
+        // Policy: Llama al método 'view' de la PermissionPolicy
+        $this->authorize('view', $permission);
+
         return PermissionResource::make($permission)->response()->setStatusCode(200);
     }
 
     public function store(PermissionRequest $request)
     {
+        // Policy: Llama al método 'create' de la PermissionPolicy
+        $this->authorize('create', Permission::class);
+
         $permission = Permission::create(['name' => $request->name]);
 
         return PermissionResource::make($permission)->response()->setStatusCode(201);
@@ -30,6 +39,9 @@ class PermissionController extends Controller
 
     public function update(PermissionRequest $request, Permission $permission)
     {
+        // Policy: Llama al método 'update' de la PermissionPolicy
+        $this->authorize('update', $permission);
+
         $permission->update(['name' => $request->name]);
 
         return PermissionResource::make($permission)->response()->setStatusCode(200);
@@ -37,6 +49,9 @@ class PermissionController extends Controller
 
     public function destroy(Permission $permission)
     {
+        // Policy: Llama al método 'delete' de la PermissionPolicy
+        $this->authorize('delete', $permission);
+
         $permission->delete();
 
         return PermissionResource::make($permission)->response()->setStatusCode(204);
