+++ app/Http/Controllers/RoleController.php
@@ -10,6 +10,9 @@ class RoleController extends Controller
 {
     public function index()
     {
+        // Policy: Llama al método 'viewAny' de la PermissionPolicy
+        $this->authorize('viewAny', Role::class);
+
         $roles = Role::all();
 
         return RoleResource::collection($roles)->response()->setStatusCode(200);
@@ -17,11 +20,17 @@ class RoleController extends Controller
 
     public function show(Role $role)
     {
+        // Policy: Llama al método 'view' de la PermissionPolicy
+        $this->authorize('view', $role);
+
         return RoleResource::make($role)->response()->setStatusCode(200);
     }
 
     public function store(RoleRequest $request)
     {
+        // Policy: Llama al método 'create' de la PermissionPolicy
+        $this->authorize('create', Role::class);
+
         // Obtenemos los permisos
         $permissions = $request->permissions ?? [];
 
@@ -35,6 +44,9 @@ class RoleController extends Controller
 
     public function update(RoleRequest $request, Role $role)
     {
+        // Policy: Llama al método 'update' de la PermissionPolicy
+        $this->authorize('update', $role);
+
         // Obtenemos los permisos
         $permissions = $request->permissions ?? [];
 
@@ -48,6 +60,9 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        // Policy: Llama al método 'delete' de la PermissionPolicy
+        $this->authorize('delete', $role);
+
         $role->delete();
 
         return RoleResource::make($role)->response()->setStatusCode(204);
